+++ how-to/workspace-platform-starter/client/types/module/shapes/platform-shapes.d.ts
@@ -27,4 +27,20 @@ export interface PlatformProviderOptions {
 	 * settings
 	 */
 	intentPicker?: IntentResolverOptions;
+	/**
+	 * When storing page/workspace data using endpoints disable the mapping which reduces the payload size.
+	 */
+	disableStorageMapping?: boolean;
+}
+/**
+ * The metadata sent with platform storage endpoint requests.
+ */
+export interface PlatformStorageMetadata {
+	/**
+	 * The version information for the storage payload.
+	 */
+	version: {
+		workspacePlatformClient: string | undefined;
+		workspaceClient: string | undefined;
+	};
 }